♻️ [Refactoring]: plugin::loader から cleanup を分離
`plugin::loader` にマニフェスト読み込みと空ディレクトリ整理が同居していたため、
責務ごとに `plugin::loader`（`load_plugin` のみ）と `plugin::cleanup`
（`cleanup_plugin_directories` と関連ヘルパー）に分割する。

- `src/plugin/cleanup.rs` を新設し cleanup 実装とテストフックを移動
- `src/plugin/loader_test.rs` → `src/plugin/cleanup_test.rs` にリネーム
  （既存テストはすべて cleanup 対象だったため）
- `src/plugin/loader.rs` は `load_plugin` のみ残し doc コメントを調整
- `src/plugin.rs` の `mod` 宣言と `pub(crate) use` を追従
- `src/target/paths.rs` のコメント内参照を新パスに更新
