+++ src/plugin.rs
@@ -1,6 +1,7 @@
 mod action;
 mod cache;
 mod cached_package;
+mod cleanup;
 mod installed;
 mod intent;
 mod loader;
@@ -15,9 +16,10 @@ mod version;
 pub use action::PluginAction;
 pub(crate) use cache::list_installed;
 pub use cache::{CachedPackage, PackageCache, PackageCacheAccess};
+pub(crate) use cleanup::cleanup_plugin_directories;
 pub use installed::InstalledPlugin;
 pub use intent::PluginIntent;
-pub(crate) use loader::{cleanup_plugin_directories, load_plugin};
+pub(crate) use loader::load_plugin;
 pub use manifest::{Author, PluginManifest};
 
 /// id フォールバック: id が None なら name を返す
