+++ src/plugin/cleanup_test.rs
